Move SessionMessage to Framework, clean up index.php/parable
