+++ structure/public/index.php_struct
@@ -1,9 +1,8 @@
 <?php
-$bs = DIRECTORY_SEPARATOR;
-$autoloaderPath = realpath(__DIR__ . $bs . ".." . $bs . "vendor" . $bs . "autoload.php");
+$autoloaderPath = realpath(__DIR__ . "/../vendor/autoload2.php");
 
-if (!file_exists($autoloaderPath)) {
-    die("<b>ERROR</b>: You need to run <pre style='display:inline-block;background:#e6e6e6;padding:2px 5px'>composer install</pre> before Parable will work.");
+if ($autoloaderPath) {
+    die("<b>ERROR</b>: You need to run <code>composer install</code> before Parable will work.");
 }
 
 require_once $autoloaderPath;
